Solved bug in checking if our king is check that was caused by not ensuring the enemy was oriented correctly. Added logic to save model, optimizer, and logs. About to run full test version of program overnight to see results.

+++ include/State.hpp
@@ -27,7 +27,7 @@ namespace Chess {
         unsigned repeated_state    : 2;  // 2 bits: 00 = first occurrence, 01 = second, 10 = third occurrence.
         unsigned half_move_count   : 6;  // 6 bits: count for the fifty-move rule (0–63).
         unsigned no_progress_side  : 1;  // 1 bit: indicates which side last made a pawn/capture move (default 0 for white).
-        uint8_t total_move_count;  // 8 bits: counts complete moves (0–255).
+        int total_move_count;  // 8 bits: counts complete moves (0–255).
     };
 
     // History snapshot to be provided to the model: just the bitboards and the repeated_state flag.
